SW-3920 Separate Tracking v2/v2 code paths for  Nursery Withdrawals
- added an initial empty placeholder component
- renamed old file to NurseryWithdrawalsV1
